pyproject.toml
[tool.poetry]
name = "endorama.github.io"
version = "0.1.0"
description = ""
authors = ["Edoardo Tenani <edoardo.tenani@pm.me>"]
license = "All Rights Reserved"

[tool.poetry.dependencies]
python = "^3.8"
proselint = "^0.13.0"
python-frontmatter = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
